% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analitica-package.R
\name{Analitica}
\alias{Analitica}
\title{Analitica: Herramientas de Análisis y Visualización de Datos}
\description{
Este paquete proporciona funciones para análisis estadístico descriptivo y visualización de variables numéricas, con y sin variables independientes.
}
\section{Funciones principales}{

- \code{\link{descripYG}}: Análisis descriptivo de variables con visualización.
}

\author{
Carlos Jimenez-Gallardo
}
\keyword{internal}
