% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_e.R
\docType{data}
\name{d_e}
\alias{d_e}
\title{Datos de clientes ficticios}
\format{
Un data.frame con N filas y M columnas. Las variables típicas pueden incluir:
\describe{
  \item{id}{ID del empleado}
  \item{Sexo}{Sexo del empleado}
  \item{FechaNac}{Fecha Nacimiento}
  \item{educacion}{cantidad de años de estudio}
  \item{labor}{area de trabajo dentro de la emrpesa}
  \item{Sueldo_actual}{sueldo a la fecha}
  \item{Sueldo_inicial}{sueldo al ingresar a la empresa}
  \item{antiguedad}{meses trabajando en la empresa}
  \item{experiencia}{meses de experiencia}
  \item{ingreso}{Ingreso mensual estimado}
  \item{minoria}{la pertenencia a una minoria}
}
}
\usage{
data(d_e)
}
\description{
Conjunto de dato, para ser utilizados como ejemplo. Las variables son:
}
\keyword{datasets}
