% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descripYG.R
\name{descripYG}
\alias{descripYG}
\title{Análisis descriptivo para una variable dependiente con/sin variable independiente}
\usage{
descripYG(dataset, vd, vi = NULL)
}
\arguments{
\item{dataset}{data.frame que contiene los datos.}

\item{vd}{Variable dependiente (numérica).}

\item{vi}{Variable independiente (opcional, categórica o continua).}
}
\value{
Un data.frame con estadísticas descriptivas. También imprime gráficos.
}
\description{
Esta función realiza un análisis descriptivo univariado o agrupado según la presencia de una variable independiente.
Incluye estadísticos como media, desviación estándar, curtosis y asimetría, y visualizaciones como histogramas, boxplots y densidades.
}
\examples{
data(iris)
descripYG(iris, vd = Sepal.Length)
descripYG(iris, vd = Sepal.Length, vi = Species)
}
