% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparations.R
\name{plot.comparaciones}
\alias{plot.comparaciones}
\title{Generic plot for multiple-comparison tests (with multcompView letters)}
\usage{
\method{plot}{comparaciones}(
  x,
  alpha = 0.05,
  p_column = c("auto", "p_ajustada", "p_value", "p"),
  horizontal = FALSE,
  fill = "steelblue",
  label_size = 5,
  angle_x = 45,
  show_se = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{comparaciones} (includes "bonferroni" or "tukey").}

\item{alpha}{Significance threshold for the letters (default 0.05).}

\item{p_column}{Which p-value column to use: "auto" (detect), "p_ajustada", "p_value", or "p".}

\item{horizontal}{If TRUE, draw horizontal bars.}

\item{fill}{Bar fill color.}

\item{label_size}{Letter size.}

\item{angle_x}{Angle of x-axis labels (if \code{horizontal = FALSE}).}

\item{show_se}{If TRUE and \code{x$MSerror} and \code{x$N} exist, draws standard error (SE) bars.}

\item{...}{Not used.}
}
\value{
A \code{ggplot} object.
}
\description{
Generic plot for multiple-comparison tests (with multcompView letters)
}
