% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparations.R
\name{plot.comparaciones}
\alias{plot.comparaciones}
\title{Generic plot for multiple comparison tests (with multcompView letters)}
\usage{
\method{plot}{comparaciones}(x, ...)
}
\arguments{
\item{x}{An object of class \code{comparaciones}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
No return value. Called for side effects: displays a bar plot with significance letters.
}
\description{
This function generates a bar plot displaying group means along with significance letters
based on multiple comparisons. It uses \code{multcompView} to assign letters indicating
statistically different groups.
}
\examples{
# Assuming you have an object of class 'comparaciones' named res
# plot(res)

}
\seealso{
\code{\link[multcompView]{multcompLetters}}, \code{\link[ggplot2]{ggplot}}
}
